requirement.txt
# Core LLM / Agent Frameworks
google-genai==1.53.0
google-adk==1.19.0
a2a-sdk==0.3.0

# MCP Server Framework
fastmcp==2.13.3

# Web + HTTP
httpx==0.28.1
uvicorn==0.38.0
aiohttp==3.13.2
anyio==4.11.0

# Async Utilities
nest_asyncio==1.6.0

# Environment Variables
python-dotenv==1.2.1

# Utilities
requests==2.32.5
